requirements.txt
python-Levenshtein
infelct